@@ -1,1 +1,1 @@
-2.6.5
+ruby-3.4.1
