@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require File.join(File.dirname(__FILE__), '..', 'gopher2000')
 
 module Gopher
@@ -6,18 +8,25 @@ module Gopher
   # @see the examples/ directory for working scripts
   #
   module DSL
-    @application = nil
-
     #
     # initialize an instance of an Application if we haven't already, otherwise, return
     #   the current app
     # @return [Gopher::Application] current app
     #
     def application
-      return @application unless @application.nil?
+      return Gopher._application unless Gopher._application.nil?
 
-      @application = Gopher::Application.new
-      @application.reset!
+      Gopher._application = Gopher::Application.new
+      Gopher._application.reset!
+    end
+
+    def application=(app)
+      if app.is_a?(Gopher::Application)
+        # @todo properly test this
+        Gopher._application = app
+      else
+        load app
+      end
     end
 
     # set a config value
@@ -30,13 +39,13 @@ module Gopher
     # specify a route
     # @param [String] path path of the route
     # @yield block that will respond to the request
-    def route(path, &block)
-      application.route(path, &block)
+    def route(path, &)
+      application.route(path, &)
     end
 
     # specify a default route
-    def default_route(&block)
-      application.default_route(&block)
+    def default_route(&)
+      application.default_route(&)
     end
 
     # mount a folder for browsing
@@ -53,31 +62,31 @@ module Gopher
         opts = opts.merge(other_opts)
       end
 
-      application.mount(route, opts.merge({:path => folder}))
+      application.mount(route, opts.merge({ path: folder }))
     end
 
     # specify a menu template
     # @param [Symbol] name of the template
     # @yield block which renders the template
-    def menu(name, &block)
-      application.menu(name, &block)
+    def menu(name, &)
+      application.menu(name, &)
     end
 
     # specify a text template
     # @param [Symbol] name of the template
     # @yield block which renders the template
-    def text(name, &block)
-      application.text(name, &block)
+    def text(name, &)
+      application.text(name, &)
     end
 
-#    def template(name, &block)
-#      application.template(name, &block)
-#    end
+    #    def template(name, &block)
+    #      application.template(name, &block)
+    #    end
 
     # specify some helpers for your app
     # @yield block which defines the helper methods
-    def helpers(&block)
-      application.helpers(&block)
+    def helpers(&)
+      application.helpers(&)
     end
 
     # watch the specified script for changes
@@ -95,22 +104,20 @@ module Gopher
     #   run on a different host/port, etc.
     #
     def run(script, opts = {})
-
       load script
 
       #
       # apply options after loading the script so that anything specified on the command-line
       # will take precedence over defaults specified in the script
       #
-      opts.each { |k, v|
+      opts.each do |k, v|
         set k, v
-      }
-
-      if application.config[:debug] == true
-        puts "watching #{script} for changes"
-        watch script
       end
 
+      return unless application.config[:debug] == true
+
+      puts "watching #{script} for changes"
+      watch script
     end
   end
 end
@@ -122,7 +129,7 @@ include Gopher::DSL
 #
 unless ENV['gopher_test']
   at_exit do
-    s = Gopher::Server.new(@application)
-    s.run!
+    s = Gopher::Server.new(Gopher._application)
+    s.run!(background: false)
   end
 end
