@@ -1,6 +1,7 @@
+# frozen_string_literal: true
+
 module Gopher
   module Rendering
-
     #
     # class for text output
     #
