@@ -1,31 +1,39 @@
-module Gopher
+# frozen_string_literal: true
 
+require 'nio'
+require 'socket'
+
+module Gopher
   #
   # main class which will listen on a specified port, and pass requests to an Application class
   #
   class Server
-    attr_accessor :app
+    attr_accessor :app, :exit
+    attr_reader :status
 
     #
     # constructor
     # @param [Application] a instance of Gopher::Application we want to run
     #
-    def initialize(a)
+    def initialize(a, host: nil, port: nil)
       @app = a
+      @status = :new
+      @host = host
+      @port = port
     end
- 
+
     #
     # @return [String] name of the host specified in our config
     #
     def host
-      @app.config[:host] ||= '0.0.0.0'
+      @host ||= @app.config[:host] ||= '0.0.0.0'
     end
 
     #
     # @return [Integer] port specified in our config
     #
     def port
-      @app.config[:port] ||= 70
+      @port ||= @app.config[:port] ||= 70
     end
 
     #
@@ -34,59 +42,98 @@ module Gopher
     def env
       @app.config[:env] || 'development'
     end
-    
+
+    def logger
+      @app.logger
+    end
+
+    def accept
+      logger.debug "Accepting new connection!"
+      socket = @server.accept
+      _, port, host = socket.peeraddr
+      logger.debug "#{host}:#{port} connected"
+
+      monitor = @selector.register(socket, :r)
+      monitor.value = proc { read(socket) }
+    end
+
+    def read(socket)
+      logger.debug "Reading from socket!"
+      _, port, host = socket.peeraddr
+
+      Dispatcher.new(app, socket).read!
+    rescue EOFError
+      logger.debug "#{host}:#{port} disconnected"
+    ensure
+      logger.debug "#{host}:#{port} cleaning up"
+      @selector.deregister(socket)
+      socket.close
+    end
+
     #
     # main app loop. called via at_exit block defined in DSL
     #
-    def run!
-      EventMachine::run do
-        Signal.trap("INT") {
-          puts "It's a trap!"
-          EventMachine.stop
-        }
-        Signal.trap("TERM") {
-          puts "It's a trap!"
-          EventMachine.stop
-        }
-
-        EventMachine.kqueue = true if EventMachine.kqueue?
-        EventMachine.epoll = true if EventMachine.epoll?
-
-
-        STDERR.puts "start server at #{host} #{port}"
-        if @app.non_blocking?
-          STDERR.puts "Not blocking on requests"
-        end
-
-
-        EventMachine::start_server(host, port, Gopher::Dispatcher) do |conn|
-          #
-          # check if we should reload any scripts before moving along
-          #
-          @app.reload_stale
-
-          #
-          # roughly matching sinatra's style of duping the app to respond
-          # to requests, @see http://www.sinatrarb.com/intro#Request/Instance%20Scope
-          #
-          # this essentially means we have 'one instance per request'
-          #
-          conn.app = @app.dup
-        end
+    def run!(background: true)
+      if background
+        @thread = Thread.new { run_server }
+      else
+        run_server
       end
     end
 
+    def stop
+      @status = :stop
+    end
+
+    def run_server
+      @selector = NIO::Selector.new
+      logger.warn "Listening on #{host}:#{port}"
+      if @app.non_blocking?
+        logger.warn "Not blocking on requests"
+      end
+
+      @server = TCPServer.new(host, port)
+
+      accept_monitor = @selector.register(@server, :r)
+      accept_monitor.value = proc { accept }
+
+      @status = :run
+
+      Signal.trap('INT') do
+        warn "It's a trap!"
+        exit! if @status == :stop
+        @status = :stop
+      end
+      Signal.trap('TERM') do
+        warn "It's a trap!"
+        @status = :stop
+      end
+
+      logger.debug "Entering loop"
+      while @status == :run
+        #
+        # check if we should reload any scripts before moving along
+        #
+        @app.reload_stale
+
+        @selector.select(1) { |monitor| monitor.value.call }
+      end
+      logger.debug "Exiting loop"
+      
+      @server.close
+      @status = :stopped
+    end
 
     #
     # don't try and parse arguments if someone already has done that
     #
-    if ARGV.any? && ! defined?(OptionParser)
+    if ARGV.any? && !defined?(OptionParser)
       require 'optparse'
-      OptionParser.new { |op|
-        op.on('-p port',   'set the port (default is 70)')                { |val| set :port, Integer(val) }
+      OptionParser.new do |op|
+        op.on('-p port',   'set the port (default is 70)') { |val| set :port, Integer(val) }
         op.on('-o addr',   'set the host (default is 0.0.0.0)')             { |val| set :host, val }
         op.on('-e env',    'set the environment (default is development)')  { |val| set :env, val.to_sym }
-      }.parse!(ARGV.dup)
+      end.parse!(ARGV.dup)
     end
   end
 end
