@@ -1,4 +1,6 @@
+# frozen_string_literal: true
+
 module Gopher
   # current version of the app
-  VERSION = "0.7.0"
+  VERSION = '0.7.0'
 end
