@@ -1,16 +1,18 @@
+# frozen_string_literal: true
+
 require File.join(File.dirname(__FILE__), '/spec_helper')
 
 describe Gopher::Application do
-  before(:each) do
-    @obj = Gopher::Application.new
-#    @obj.extend(Gopher::Helpers)
+  before do
+    @obj = described_class.new
+    #    @obj.extend(Gopher::Helpers)
   end
 
-  it 'should add code to target class' do
+  it 'adds code to target class' do
     @obj.helpers do
-      def foo; "FOO"; end
+      def foo = 'FOO'
     end
 
-    expect(@obj.foo).to eq("FOO")
+    expect(@obj.foo).to eq('FOO')
   end
 end
